backend/requirements.txt
# FastAPI ve temel web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Veritabanı
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis ve Celery
redis==5.0.1
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI ve görsel işleme
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
ultralytics==8.0.196
mediapipe==0.10.7
rembg==2.0.50

# Utility
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 